Move load_fasta_header into utils
