+++ thapbi_pict/utils.py
@@ -779,3 +779,27 @@ def color_bands(meta_groups, sample_color_bands, debug=False):
             debug_msg.append(value)
     assert len(set(bands)) == max(bands) + 1, bands
     return [sample_color_bands[_ % len(sample_color_bands)] for _ in bands]
+
+
+def load_fasta_header(fasta_file, gzipped=False):
+    """Parse our FASTA hash-comment line header as a dict."""
+    answer = {}
+    if gzipped:
+        handle = gzip.open(fasta_file, "rt")
+    else:
+        handle = open(fasta_file)
+    for line in handle:
+        if line.startswith("#") and ":" in line:
+            tag, value = line[1:].strip().split(":", 1)
+            try:
+                value = int(value)
+            except ValueError:
+                pass
+            answer[tag] = value
+        elif line.startswith(">"):
+            break
+        elif not line.strip():
+            pass
+        else:
+            sys.exit(f"ERROR: Unexpected line in headered FASTA file {fasta_file}")
+    return answer
